src/workshop/requirements.txt
aiosqlite
httpx
aiohttp
python_dotenv
azure-identity
azure-ai-projects
azure-ai-agents
pandas
pydantic
pillow